Moved save basis code to wrapper
